pyproject.toml
[project]
name = "dhex"
version = "0.1.0"
description = "DH Exchange API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio>=0.23.6",
]
